Cleanup applications after adding assessment/analysis tabs
