+++ pkg/client/public/locales/es/translation.json
@@ -80,7 +80,6 @@
             "discard": "¿Descartar {{what}}?",
             "importApplicationFile": "Importar archivo con aplicaciones",
             "leavePage": "Salir de página",
-            "newApplication": "Nueva aplicación",
             "newBusinessService": "Nuevo servicio de negocio",
             "newJobFunction": "Nueva función laboral",
             "newStakeholder": "Nuevo interesado",
