+++ pkg/client/src/app/pages/applications/applications-table-analyze/index.ts
@@ -1,0 +1,1 @@
+export { ApplicationsTableAnalyze as default } from "./applications-table-analyze";
