+++ pkg/client/src/app/pages/applications/applications-table-assessment/applications-table-assessment.tsx
@@ -76,13 +76,14 @@ import {
 import { applicationPageMapper } from "@app/api/apiUtils";
 import { getAxiosErrorMessage } from "@app/utils/utils";
 
-import { ApplicationForm } from "./components/application-form";
+import { ApplicationForm } from "../components/application-form";
 
-import { ApplicationBusinessService } from "./components/application-business-service";
-import { ApplicationListExpandedArea } from "./components/application-list-expanded-area";
-import { ImportApplicationsForm } from "./components/import-applications-form";
-import { BulkCopyAssessmentReviewForm } from "./components/bulk-copy-assessment-review-form";
-import { ApplicationsIdentityForm } from "./components/ApplicationsIdentityForm";
+import { ApplicationAssessment } from "../components/application-assessment";
+import { ApplicationBusinessService } from "../components/application-business-service";
+import { ApplicationListExpandedArea } from "../components/application-list-expanded-area";
+import { ImportApplicationsForm } from "../components/import-applications-form";
+import { BulkCopyAssessmentReviewForm } from "../components/bulk-copy-assessment-review-form";
+import { ApplicationsIdentityForm } from "../components/ApplicationsIdentityForm";
 
 const toSortByQuery = (
   sortBy?: SortByQuery
@@ -125,7 +126,7 @@ const searchAppAssessment = (id: number) => {
   return result;
 };
 
-export const ApplicationsTableAnalyze: React.FC = () => {
+export const ApplicationsTable: React.FC = () => {
   // i18
   const { t } = useTranslation();
 
@@ -216,6 +217,7 @@ export const ApplicationsTableAnalyze: React.FC = () => {
     if (!applicationToUpdate) {
       dispatch(
         alertActions.addSuccess(
+          // t('terms.application')
           t("toastr.success.added", {
             what: response.data.name,
             type: t("terms.application").toLowerCase(),
@@ -313,7 +315,8 @@ export const ApplicationsTableAnalyze: React.FC = () => {
     },
     { title: t("terms.description"), transforms: [cellWidth(25)] },
     { title: t("terms.businessService"), transforms: [cellWidth(20)] },
-    { title: t("terms.analysis"), transforms: [sortable, cellWidth(10)] },
+    { title: t("terms.assessment"), transforms: [cellWidth(10)] },
+    { title: t("terms.review"), transforms: [sortable, cellWidth(10)] },
     { title: t("terms.tagCount"), transforms: [sortable, cellWidth(10)] },
     {
       title: "",
@@ -351,8 +354,16 @@ export const ApplicationsTableAnalyze: React.FC = () => {
           ),
         },
         {
-          // "Not started" | "Scheduled" | "In-progress" | "Canceled" | "Failed" | "Completed"
-
+          title: (
+            <ApplicationAssessment
+              assessment={getApplicationAssessment(item.id!)}
+              isFetching={isFetchingApplicationAssessment(item.id!)}
+              fetchError={fetchErrorApplicationAssessment(item.id!)}
+              fetchCount={fetchCountApplicationAssessment(item.id!)}
+            />
+          ),
+        },
+        {
           title: item.review ? (
             <StatusIconAssessment status="Completed" />
           ) : (
@@ -603,7 +614,6 @@ export const ApplicationsTableAnalyze: React.FC = () => {
         if (assessment.status === "COMPLETE") {
           dispatch(
             confirmDialogActions.openDialog({
-              // t("terms.assessment")
               title: t("composed.editQuestion", {
                 what: t("terms.assessment").toLowerCase(),
               }),
@@ -701,7 +711,7 @@ export const ApplicationsTableAnalyze: React.FC = () => {
                 <ToolbarItem>
                   <Button
                     type="button"
-                    aria-label="analyze-application"
+                    aria-label="assess-application"
                     variant={ButtonVariant.primary}
                     onClick={assessSelectedRows}
                     isDisabled={
@@ -709,7 +719,21 @@ export const ApplicationsTableAnalyze: React.FC = () => {
                     }
                     isLoading={isApplicationAssessInProgress}
                   >
-                    {t("actions.analyze")}
+                    {t("actions.assess")}
+                  </Button>
+                </ToolbarItem>
+                <ToolbarItem>
+                  <Button
+                    type="button"
+                    aria-label="review-application"
+                    variant={ButtonVariant.primary}
+                    onClick={reviewSelectedRows}
+                    isDisabled={
+                      selectedRows.length !== 1 ||
+                      isReviewBtnDisabled(selectedRows[0])
+                    }
+                  >
+                    {t("actions.review")}
                   </Button>
                 </ToolbarItem>
                 <ToolbarItem>
