+++ pkg/client/src/app/pages/applications/applications-table-assessment/index.ts
@@ -1,0 +1,1 @@
+export { ApplicationsTable as default } from "./applications-table-assessment";
