--- pkg/client/src/app/pages/applications/applicationsTable/components/new-application-modal/index.ts
@@ -1,1 +1,0 @@
-export { NewApplicationModal } from "./new-application-modal";
