--- pkg/client/src/app/pages/applications/applicationsTable/index.ts
@@ -1,1 +1,0 @@
-export { ApplicationsTable as default } from "./ApplicationsTable";
