--- pkg/client/src/app/pages/applications/applicationsTableAnalyze/components/application-assessment/index.ts
@@ -1,1 +1,0 @@
-export { ApplicationAssessment } from "./application-assessment";
