--- pkg/client/src/app/pages/applications/applicationsTableAnalyze/components/application-business-service/index.ts
@@ -1,1 +1,0 @@
-export { ApplicationBusinessService } from "./application-business-service";
