--- pkg/client/src/app/pages/applications/applicationsTableAnalyze/components/application-form/application-form.css
@@ -1,4 +1,0 @@
-.toggle.pf-c-expandable-section {
-  --pf-c-expandable-section__toggle--Color: #000;
-  --pf-c-expandable-section__toggle--m-expanded--Color: #000;
-}
