--- pkg/client/src/app/pages/applications/applicationsTableAnalyze/components/application-form/index.ts
@@ -1,1 +1,0 @@
-export { ApplicationForm } from "./application-form";
