--- pkg/client/src/app/pages/applications/applicationsTableAnalyze/components/application-list-expanded-area/index.ts
@@ -1,1 +1,0 @@
-export { ApplicationListExpandedArea } from "./application-list-expanded-area";
