--- pkg/client/src/app/pages/applications/applicationsTableAnalyze/components/application-tags/index.ts
@@ -1,1 +1,0 @@
-export { ApplicationTags } from "./application-tags";
