--- pkg/client/src/app/pages/applications/applicationsTableAnalyze/components/bulk-copy-assessment-review-form/index.tsx
@@ -1,1 +1,0 @@
-export { BulkCopyAssessmentReviewForm } from "./bulk-copy-assessment-review-form";
