--- pkg/client/src/app/pages/applications/applicationsTableAnalyze/components/import-applications-form/index.ts
@@ -1,1 +1,0 @@
-export { ImportApplicationsForm } from "./import-applications-form";
