--- pkg/client/src/app/pages/applications/applicationsTableAnalyze/components/update-application-modal/index.ts
@@ -1,1 +1,0 @@
-export { UpdateApplicationModal } from "./update-application-modal";
