--- pkg/client/src/app/pages/applications/applicationsTableAnalyze/index.ts
@@ -1,1 +1,0 @@
-export { ApplicationsTableAnalyze as default } from "./ApplicationsTableAnalyze";
