+++ pkg/client/src/app/pages/applications/components/application-form/application-form.css
