+++ pkg/client/src/app/pages/applications/components/application-list-expanded-area/application-list-expanded-area.tsx
