+++ pkg/client/src/app/pages/applications/components/application-list-expanded-area/index.ts
