+++ pkg/client/src/app/pages/applications/components/application-list-expanded-area/tests/application-list-expanded-area.test.tsx
