+++ pkg/client/src/app/pages/applications/components/application-tags/application-tags.tsx
