+++ pkg/client/src/app/pages/applications/components/application-tags/index.ts
